requirements.txt
Appium-Python-Client==2.11.0
pytest==8.1.1
pytest-html==4.0.2
